requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
google-api-python-client==2.147.0
google-auth==2.33.0
google-auth-oauthlib==1.2.1
openai==1.42.0
aiofiles==24.1.0
gunicorn==22.0.0
gTTS==2.5.3
openai-whisper==20231117
